-Moved all mobile dev dev data files into their respective folder -Deleted old data files -Imported new data file outlines into the job explorer slug -created files for the mobile dev project data
