+++ pages/api/resources/jobExplorer/androidDev/data.ts
@@ -1,0 +1,26 @@
+export const ANDROID_DEVELOPER_DATA = {
+    title: "Android Developer",
+    description: "description",
+    mandatorySkills: [
+      {
+        name: "skill1",
+        link: "link",
+      },
+      {
+        name: "skill2",
+        link: "link",
+      },
+    ],
+    additionalSkills: [
+      {
+        name: "skill1",
+        link: "link",
+      },
+      {
+        name: "skill2",
+        link: "",
+      },
+    ],
+    project: { name: "project", link: "link" },
+  };
+
