--- pages/api/resources/jobExplorer/androidDeveloper.ts
@@ -1,25 +1,0 @@
-export const ANDROID_DEVELOPER_DATA = {
-  title: "Android Developer",
-  description: "description",
-  mandatorySkills: [
-    {
-      name: "skill1",
-      link: "link",
-    },
-    {
-      name: "skill2",
-      link: "link",
-    },
-  ],
-  additionalSkills: [
-    {
-      name: "skill1",
-      link: "link",
-    },
-    {
-      name: "skill2",
-      link: "",
-    },
-  ],
-  project: { name: "project", link: "link" },
-};
