+++ pages/api/resources/jobExplorer/iosDev/data.ts
@@ -1,0 +1,26 @@
+export const IOS_DEVELOPER_DATA = {
+    title: "IOS Developer",
+    description: "description",
+    mandatorySkills: [
+      {
+        name: "skill1",
+        link: "link",
+      },
+      {
+        name: "skill2",
+        link: "link",
+      },
+    ],
+    additionalSkills: [
+      {
+        name: "skill1",
+        link: "link",
+      },
+      {
+        name: "skill2",
+        link: "",
+      },
+    ],
+    project: { name: "project", link: "link" },
+  };
+
