--- pages/api/resources/jobExplorer/iosDeveloper.ts
@@ -1,25 +1,0 @@
-export const IOS_DEVELOPER_DATA = {
-  title: "IOS Developer",
-  description: "description",
-  mandatorySkills: [
-    {
-      name: "skill1",
-      link: "link",
-    },
-    {
-      name: "skill2",
-      link: "link",
-    },
-  ],
-  additionalSkills: [
-    {
-      name: "skill1",
-      link: "link",
-    },
-    {
-      name: "skill2",
-      link: "",
-    },
-  ],
-  project: { name: "project", link: "link" },
-};
