--- pages/api/resources/jobExplorer/mobileAppDev/data.ts
@@ -1,87 +1,0 @@
-export const title = "Mobile App Developer";
-
-export const description =
-  "A mobile app developer is a skilled software engineer who specializes in designing, developing, and maintaining mobile applications for smartphones and other mobile devices. These developers play a crucial role in the ever-expanding world of mobile technology, creating innovative and user-friendly apps that run on various platforms like Android, iOS, or cross-platform frameworks.";
-
-export const mandatorySkills = [
-  {
-    name: "Android",
-    link: "https://developer.android.com/",
-  },
-  {
-    name: "iOS",
-    link: "https://developer.apple.com/ios/",
-  },
-  {
-    name: "Java",
-    link: "https://docs.oracle.com/en/java/javase/index.html",
-  },
-  {
-    name: "Swift",
-    link: "https://developer.apple.com/swift/",
-  },
-  {
-    name: "Mobile User Interface (UI) Design",
-    link: "https://developer.apple.com/design/human-interface-guidelines/",
-  },
-  {
-    name: "Mobile App Testing and Debugging",
-    link: "https://developer.android.com/studio/debug/",
-  },
-];
-
-export const additionalSkills = [
-  {
-    name: "Cross-platform App Development",
-    link: "https://flutter.dev/docs",
-  },
-  {
-    name: "Mobile App Performance Optimization",
-    link: "https://developer.android.com/topic/performance",
-  },
-  {
-    name: "App Security",
-    link: "https://www.owasp.org/index.php/OWASP_Mobile_Security_Project",
-  },
-  {
-    name: "RESTful API Integration",
-    link: "https://developer.mozilla.org/en-US/docs/Web/HTTP/Methods/GET",
-  },
-  {
-    name: "Push Notifications Integration",
-    link: "https://firebase.google.com/docs/cloud-messaging",
-  },
-  {
-    name: "Mobile App Analytics",
-    link: "https://firebase.google.com/docs/analytics",
-  },
-  {
-    name: "In-App Purchases and Payment Gateway Integration",
-    link: "https://developer.android.com/google/play/billing",
-  },
-  {
-    name: "Offline Data Synchronization",
-    link: "https://developer.android.com/training/sync-adapters",
-  },
-  {
-    name: "Background Processing and Multithreading",
-    link: "https://developer.android.com/guide/components/processes-and-threads",
-  },
-  {
-    name: "Mobile App Accessibility",
-    link: "https://developer.android.com/guide/topics/ui/accessibility",
-  },
-  {
-    name: "Augmented Reality (AR) or Virtual Reality (VR) Integration",
-    link: "https://developers.google.com/ar",
-  },
-  {
-    name: "Mobile App Localization (Multi-language Support)",
-    link: "https://developer.android.com/guide/topics/resources/localization",
-  },
-];
-
-export const project = {
-  name: "Mobile App Developer Project",
-  link: "mobileAppDevProject",
-};
